Split marker tests into separate file to avoid conflicts
